Fix [Metrics] UI metrics UX fix (#2552)
