+++ src/api/details-api.js
@@ -31,23 +31,9 @@ const detailsApi = {
 
     return mainHttpClient.get(`/projects/${project}/runtime-resources`, { params })
   },
-  getModelEndpoint: (project, uid) =>
-    mainHttpClient.get(`/projects/${project}/model-endpoints/${uid}?feature_analysis=true`),
+
   getModelFeatureVector: (project, name, reference) =>
-    mainHttpClient.get(`/projects/${project}/feature-vectors/${name}/references/${reference}`),
-  getModelEndpointMetrics: (project, uid, type = 'all') =>
-    mainHttpClient.get(`/projects/${project}/model-endpoints/${uid}/metrics?type=${type}`), // type=results/metrics/all
-  getModelEndpointMetricsValues: (project, uid, params, signal) => {
-    const config = {
-      params
-    }
-
-    if (signal) {
-      config.signal = signal
-    }
-
-    return mainHttpClient.get(`/projects/${project}/model-endpoints/${uid}/metrics-values`, config)
-  }
+    mainHttpClient.get(`/projects/${project}/feature-vectors/${name}/references/${reference}`)
 }
 
 export default detailsApi
