+++ src/common/TabsSlider/tabsSlider.scss
@@ -5,7 +5,7 @@
   display: flex;
   align-items: stretch;
   background-color: $white;
-  padding-bottom: 30px;
+  padding-bottom: 15px;
 
   &__tabs-wrapper {
     overflow-x: hidden;
