+++ src/components/Details/Details.js
@@ -27,11 +27,12 @@ import { Form } from 'react-final-form'
 import { isEqual, pickBy } from 'lodash'
 import classnames from 'classnames'
 
-import DetailsTabsContent from './DetailsTabsContent/DetailsTabsContent'
 import { ConfirmDialog } from 'igz-controls/components'
-import Loader from '../../common/Loader/Loader'
 import ErrorMessage from '../../common/ErrorMessage/ErrorMessage'
+import DetailsTabsContent from './DetailsTabsContent/DetailsTabsContent'
+
 import DetailsHeader from './DetailsHeader/DetailsHeader'
+import Loader from '../../common/Loader/Loader'
 import TabsSlider from '../../common/TabsSlider/TabsSlider'
 
 import { TERTIARY_BUTTON, PRIMARY_BUTTON } from 'igz-controls/constants'
@@ -55,7 +56,6 @@ import {
 import { isEveryObjectValueEmpty } from '../../utils/isEveryObjectValueEmpty'
 import { showArtifactsPreview } from '../../reducers/artifactsReducer'
 import { setFieldState } from 'igz-controls/utils/form.util'
-import { datePickerPastOptions, PAST_24_HOUR_DATE_OPTION } from '../../utils/datePicker.util'
 
 import './details.scss'
 
@@ -78,11 +78,9 @@ const Details = ({
   setChanges,
   setChangesCounter,
   setChangesData,
-  setDetailsDates,
   setInfoContent,
   setIteration,
   setIterationOption,
-  setSelectedMetricsOptions,
   setFiltersWasHandled,
   showWarning,
   tab
@@ -248,30 +246,6 @@ const Details = ({
     setFiltersWasHandled
   ])
 
-  const handleChangeDates = useCallback(
-    (dates, isPredefined, selectedOptionId) => {
-      const generatedDates = [...dates]
-
-      if (generatedDates.length === 1) {
-        generatedDates.push(new Date())
-      }
-      setDetailsDates({
-        value: generatedDates,
-        selectedOptionId,
-        isPredefined
-      })
-    },
-    [setDetailsDates]
-  )
-
-  useEffect(() => {
-    const past24hoursOption = datePickerPastOptions.find(
-      option => option.id === PAST_24_HOUR_DATE_OPTION
-    )
-
-    handleChangeDates(past24hoursOption.handler(), true, PAST_24_HOUR_DATE_OPTION)
-  }, [handleChangeDates])
-
   return (
     <Form form={formRef.current} onSubmit={() => {}}>
       {formState => (
@@ -287,13 +261,11 @@ const Details = ({
               getCloseDetailsLink={getCloseDetailsLink}
               isDetailsScreen={isDetailsScreen}
               handleCancel={handleCancel}
-              handleChangeDates={handleChangeDates}
               handleRefresh={handleRefresh}
               handleShowWarning={handleShowWarning}
               pageData={pageData}
               selectedItem={selectedItem}
               setIteration={setIteration}
-              setSelectedMetricsOptions={setSelectedMetricsOptions}
               tab={tab}
             />
             <TabsSlider tabsList={detailsMenu} initialTab={params.tab} />
