+++ src/components/Details/details.scss
@@ -66,7 +66,6 @@
 
         &__navigation-buttons {
           display: flex;
-          margin-left: 20px;
         }
 
         &__pods {
