+++ src/components/FunctionsPage/functions.util.js
@@ -247,18 +247,18 @@ export const generateActionsMenu = (
           !FUNCTIONS_EDITABLE_STATES.includes(func?.state?.value)
       },
       {
-        label: 'Delete',
-        icon: <Delete />,
-        className: 'danger',
-        disabled: functionIsDeleting,
-        onClick: onRemoveFunction
-      },
-      {
         label: 'View YAML',
         icon: <Yaml />,
         disabled: functionIsDeleting,
         onClick: funcMin =>
           getFullFunction(funcMin).then(func => !isEmpty(func) && toggleConvertedYaml(func))
+      },
+      {
+        label: 'Delete',
+        icon: <Delete />,
+        className: 'danger',
+        disabled: functionIsDeleting,
+        onClick: onRemoveFunction
       }
     ],
     [
