+++ src/components/ProjectsJobsMonitoring/projectsJobsMonitoring.scss
@@ -1,8 +1,4 @@
 .job-monitoring {
-  .content__action-bar-wrapper {
-    margin-bottom: 25px;
-  }
-
   .table-top {
     margin-top: 0;
     margin-bottom: 20px;
