+++ src/elements/MetricsSelector/metricsSelector.scss
@@ -2,8 +2,6 @@
 @import '~igz-controls/scss/borders';
 
 .metrics-selector {
-  padding-right: 16px;
-
   &-header {
     position: relative;
     display: flex;
