+++ src/scss/main.scss
@@ -94,7 +94,7 @@ body {
   &__action-bar-wrapper {
     display: flex;
     align-items: center;
-    padding: 4px 0;
+    padding: 4px 0 15px 0;
 
     &_hidden {
       display: none;
@@ -102,7 +102,7 @@ body {
 
     .action-bar {
       display: flex;
-      align-items: center;
+      align-items: flex-start;
       justify-content: flex-end;
       width: 100%;
 
@@ -795,8 +795,6 @@ main {
 }
 
 .action-bar {
-  padding: 20px 0;
-
   .filters-button {
     margin-left: 5px;
   }
